Set up some Mesh Stuff
Although the CubeMesh class is there, no cube data is properly inputed. @TODO tommorow

+++ Raytracer/Vector3.java
@@ -9,7 +9,7 @@ public class Vector3
     public float x;
     public float y;
     public float z;
-    public Vector3(int x, int y, int z)
+    public Vector3(float x, float y, float z)
     {
         this.x = x;
         this.y = y;
